mesh: update torch
