+++ fealpy/mesh/torch/mesh_data_structure/__init__.py
@@ -1,0 +1,3 @@
+from .mesh_ds import MeshDataStructure, Redirector
+from .mesh1d_ds import Mesh1dDataStructure
+from .mesh2d_ds import Mesh2dDataStructure
